on start, install the terminal shortcut scripts 'debian' and '/debian/shell' (closes #2)
these scripts make it easy to get a shell in the Debian chroot in
Terminal Emulator and programs like it, but they don't run automatically,
so they don't force the user to always work in the Debian chroot.

If someone wants to only have Debian shells in Terminal Emulator, they
can add /debian/shell in the preferences of Terminal Emulator.