use tune2fs to switch to ext3 if available after formatting new imagefile, fixes #43
It seems that the busybox mke2fs only supports formatting as ext2, and its
tune2fs does not support the -j flag for turning on the ext3 journal. So in
order to use ext3, format then convert using Debian's tune2fs. To use
ext4, it should be mountable, but it'll have to be formatted off the device