create shortcut scripts as early as possible, since they're only created here
The shortcut scripts /debian/shell and /data/local/bin/debian are only
create by the create-debian-setup.sh script. If there is an error in the
script before those shortcuts are created, you can be left with a usable
system but without those shortcuts. Create those scripts as early as
possible to they're available if there is a working setup.