run native setup in Application subclass to ensure its always run first
The android.app.Application instance associated with the app is always
launched first, so its the natural place to run the native setup.  This
fixes crashes that happen if the InstallLog was the last Activity viewed,
then the app gets killed and relaunched.  Before, it would launch and go
straight to InstallLogViewActivity and never run the native setup.

Most of the code in NativeHelper should probably be moved to
LilDebiApplication.