use faketime on release builds for deterministic builds
Running a program with faketime causes that program to recent a fixed time
based on the timestamp provided to faketime.  This ensures that the
timestamps in the files are always the same.

This gets the process to the point where the APK signature is transferrable
no matter who builds it.  This process does not produce an APK with the
same hash.