use special faketime timestamp format to freeze time
faketime needs to have a very specific timestamp format in order to freeze
time.  The time needs to be frozen so that the timestamps don't depend on
the speed of the machine that the build process is running on.  See `man
faketime` for more info on the "advanced timestamp format".  Also, force
time to UTC so its always the same on all machines.

https://lists.mayfirst.org/pipermail/guardian-dev/2014-April/003420.html