force timezone to UTC for busybox
busybox includes a timestamp in its code, set in a C macro called both
BB_BT and AUTOCONF_TIMESTAMP.  BB_BT is set to AUTOCONF_TIMESTAMP in the
build flags.  AUTOCONF_TIMESTAMP is set in busybox/include/autoconf.h. This
timestamp includes the time zone, so make sure it is always UTC.