calculating free size of internal storage from '/data' and not '/data/debian'
When lildebi is installed first time on a device, then /data/debian folder is
not present. So when the install screen (of installing in internal storage)
tries to calculate the free size of internal storage from /data/debian then an
exception is generated as /data/debian is not present.  So instead of
calculating free size from /data/debian, we will now calculate it from /data

fixes #114 https://github.com/guardianproject/lildebi/issues/114