Debian's /etc/profile needs real `id -u`, so use busybox's
Android's `id` is in the PATH when this little snipped of code is run in
Debian.  So when /etc/profile is first executed and is setting the PATH,
it needs `id -u` to return only a numeric value.

fixes #88 https://github.com/guardianproject/lildebi/issues/88
fixes #116 https://github.com/guardianproject/lildebi/issues/116