write install log directly to file; broadcast updates to InstallActivity
This commit should drastically improve the performance of the install
process.  The old process was really horribly inefficient, it reposted the
entire contents of the log in order to update a single line.  This commit
also makes it so you can follow the log file in realtime using `tail -f` or
whatever.

closes #137 https://github.com/guardianproject/lildebi/issues/137