include current Android Locale in Debian setup
This sets the language on the Debian install based on how the language is
set in Debian.  It will only generate the locales on install, but it will
always set the current locale in start-debian.sh.  If the current locale is
differen than the locale at install time, the user needs to manually gen
the current locale in Debian.

closes #140 https://github.com/guardianproject/lildebi/issues/140